backend/requirements.txt
Flask>=2.0
python-dotenv>=0.15
requests>=2.25
psycopg2-binary>=2.9 # For PostgreSQL
pytest>=7.0
pytest-cov>=4.0
